refactor upgrade utils for testing
